Separate out the delete permission UI into it's own JS file.
